splitting routes
